db_gateway/requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Cors==4.0.0
Flask-GraphQL==2.0.1
flask-sqlalchemy==3.0.5
graphql-core==2.3.2
graphql-server-core==1.2.0
greenlet==2.0.2
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
promise==2.3
PyMySQL==1.1.0
python-dotenv==1.0.0
Rx==1.6.3
six==1.16.0
SQLAlchemy==2.0.17
typing-extensions==4.7.1
Werkzeug==2.3.6
zipp==3.15.0
